pyproject.toml
[project]
name = "recycled_softening_analysis"
readme = "README.md"
dynamic = ["version"]
dependencies = [
    "pyomo","idaes-pse>=2.5.0","watertap>=1.0.0"
]
